cothon solution projecd/requirements.txt
streamlit>=1.31.1
pandas>=2.2.0
plotly>=5.18.0
nltk>=3.8.1
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0
typing-extensions>=4.9.0
python-dateutil>=2.8.2
pytz>=2024.1
numpy>=1.26.0
tomli>=2.0.1; python_version < "3.13"
